requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
opencv-python==4.10.0
numpy==2.2.1
torch==2.5.1
torchvision==0.20.1
Pillow==11.0.0
mediapipe==0.10.20

-e .

